requirements.txt
docopt
PyYAML
selenium
pyinstaller